+++ src/frontend/admin/src/components/presentational/hook-form/RHFTextField.tsx
@@ -16,11 +16,7 @@ export function RHFTextField({ name, helperText, ...props }: Props) {
         <TextField
           {...field}
           fullWidth
-          value={
-            typeof field.value === "number" && field.value === 0
-              ? ""
-              : field.value
-          }
+          value={field.value}
           error={!!error}
           helperText={error ? error?.message : helperText}
           {...props}